Extract common worker functionality to MVTWorker class
