+++ src/client/windshaft.worker.js
@@ -1,5 +1,4 @@
 import { MVTWorker } from '../sources/MVTWorker';
-import Metadata from '../renderer/Metadata';
 import schema from '../renderer/schema';
 
 class WindshaftWorker extends MVTWorker {
@@ -31,26 +30,4 @@ class WindshaftWorker extends MVTWorker {
 
 const worker = new WindshaftWorker();
 
-onmessage = function (event) {
-    processEvent(event).then(message => {
-        message.dataframe.geom = null;
-        const transferables = [];
-        if (!message.dataframe.empty) {
-            transferables.push(message.dataframe.decodedGeom.verticesArrayBuffer);
-            if (message.dataframe.decodedGeom.normalsArrayBuffer) {
-                transferables.push(message.dataframe.decodedGeom.normalsArrayBuffer);
-            }
-        }
-        postMessage(message, transferables);
-    });
-};
-
-async function processEvent (event) {
-    const params = event.data;
-    Object.setPrototypeOf(params.metadata, Metadata.prototype);
-    const dataframe = await worker._requestDataframe(params.x, params.y, params.z, params.url, params.layerID, params.metadata);
-    return {
-        mID: params.mID,
-        dataframe
-    };
-}
+onmessage = worker.onmessage.bind(worker);
